ref:clean code server
